requirements.txt
Django>=4.2
psycopg2-binary>=2.9
graphene-django>=3.2
django-graphql-jwt>=0.4
graphene-file-upload>=1.3.0 
django-cors-headers>=4.3.1 
uuid>=1.30 